requirements.txt
torch
torchvision
pyyaml
numpy
Pillow
